+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.1 - 2021-10-15
+
+### Fixed
+- The XML file format changed, so the processing step was changed accordingly.
+
+### Changed
+- Save relations even if only one gene was detected. 
+
 ## 0.1.1 - 2021-04-22
 
 ### Fixed